backend/pyproject.toml
[project]
name = "raptor-mvp"
version = "0.1.0"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "pydantic>=2.7",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "PyJWT>=2.9",
  "Pillow>=10.4",
  "ExifRead>=3.0",
  "pyproj>=3.6",
  "shapely>=2.0",
  "pyshp>=2.3",
  "docxtpl>=0.16",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
